+++ comm/commands.c
@@ -56,6 +56,7 @@
 #endif
 #include "main.h"
 #include "conf_custom.h"
+#include "comm_usb.h"
 
 #include <math.h>
 #include <string.h>
@@ -1761,6 +1762,9 @@ int commands_printf_lisp(const char* format, ...) {
 		}
 
 		commands_send_packet((unsigned char*)print_buffer, len_to_print);
+
+		// Uncomment to always print to USB. Useful when debugging code that redirects prints
+//		comm_usb_send_packet((unsigned char*)print_buffer, len_to_print);
 	}
 
 	chMtxUnlock(&print_mutex);
@@ -2248,9 +2252,12 @@ static THD_FUNCTION(blocking_thread, arg) {
 
 			float linkage, linkage_undriven, undriven_samples;
 			bool res;
+			float enc_offset, enc_ratio;
+			bool enc_inverted;
 			int fault = conf_general_measure_flux_linkage_openloop(current, duty,
 																   erpm_per_sec, resistance, inductance,
-																   &linkage, &linkage_undriven, &undriven_samples, &res);
+																   &linkage, &linkage_undriven, &undriven_samples, &res,
+																   &enc_offset, &enc_ratio, &enc_inverted);
 
 			if (fault != FAULT_CODE_NONE) {
 				linkage = 0.0;
@@ -2268,6 +2275,9 @@ static THD_FUNCTION(blocking_thread, arg) {
 			ind = 0;
 			send_buffer[ind++] = COMM_DETECT_MOTOR_FLUX_LINKAGE_OPENLOOP;
 			buffer_append_float32(send_buffer, linkage, 1e7, &ind);
+			buffer_append_float32(send_buffer, enc_offset, 1e6, &ind);
+			buffer_append_float32(send_buffer, enc_ratio, 1e6, &ind);
+			send_buffer[ind++] = enc_inverted;
 			if (send_func_blocking) {
 				send_func_blocking(send_buffer, ind);
 			}