+++ CHANGELOG.md
@@ -11,12 +11,15 @@
 		* Super fast boot possible
 		* Much easier to use const blocks
 		* Rebuild image and const data when needed
+	* Added conf-detect-lambda-enc.
+	* Added many configuration parameters.
 * New offset calibration modes and options.
 * Automatic offset calibration support.
 * Added HFI ambiguity resolution modes using id injection.
 * Support for coordinated CAN baudrate update.
 * Overmodulation support.
 * Added PWM and PWM+ABI encoder modes.
+* Detect encoder parameters during flux linkage detection.
 
 ### 6.05
 #### 2024-08-19